requirements.txt
flask
chromadb
sentence-transformers
langchain
langchain-community
langchain-text-splitters
langchain-chroma
requests
qrcode[pil]
pillow
langdetect
gTTS